requirements.txt
python-dotenv
django
celery
psycopg2-binary
requests
beautifulsoup4
redis
lxml